ci/pyproject.toml
[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "bmtgate"
module-root = "src"

[project]
name = "kardome-bmt-gate"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
description = "Kardome BMT CI driver: matrix, trigger, VM lifecycle, runner upload (import: bmtgate)."
dependencies = [
    "kardome-bmt-contract",
    "google-cloud-compute>=1.45",
    "google-cloud-pubsub>=2.21",
    "google-cloud-storage>=2.16",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0.0",
    "githubkit>=0.15.1",
]

[dependency-groups]
dev = [
    "ty",
    "ruff>=0.15.0",
]

[project.scripts]
bmt = "bmtgate.main:main"
bmt-matrix = "bmtgate.main:main_matrix"
bmt-write-context = "bmtgate.main:main_write_context"
bmt-write-summary = "bmtgate.main:main_write_handoff_summary"

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [".venv", "__pycache__", ".ruff_cache"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.ty.environment]
python-version = "3.12"
